+++ scripts/devops_tasks/tox_harness.py
@@ -19,6 +19,7 @@ from common_tasks import (
     log_file,
     read_file,
     MANAGEMENT_PACKAGE_IDENTIFIERS,
+    create_code_coverage_params,
 )
 
 import logging
@@ -74,11 +75,35 @@ class ThreadPool:
     def wait_completion(self):
         self.tasks.join()
 
+def combine_coverage_files(targeted_packages):
+    # find tox.ini file. tox.ini is used to combine coverage paths to generate formatted report
+    tox_ini_file = os.path.join(root_dir, "eng", "tox", "tox.ini")
+    config_file_flag = "--rcfile={}".format(tox_ini_file)
+
+    if os.path.isfile(tox_ini_file):
+        # for every individual coverage file, run coverage combine to combine path
+        for package_dir in [package for package in targeted_packages]:
+            coverage_file = os.path.join(package_dir, ".coverage")
+            if os.path.isfile(coverage_file):
+                cov_cmd_array = [sys.executable, "-m", "coverage", "combine"]
+                # tox.ini file has coverage paths to combine
+                # Pas tox.ini as coverage config file
+                cov_cmd_array.extend([config_file_flag, coverage_file])
+                run_check_call(cov_cmd_array, package_dir)
+    else:
+        # not a hard error at this point
+        # this combine step is required only for modules if report has package name starts with .tox
+        logging.error("tox.ini is not found in path {}".format(root_dir))
+
 
 def collect_tox_coverage_files(targeted_packages):
     root_coverage_dir = os.path.join(root_dir, "_coverage/")
 
     clean_coverage(coverage_dir)
+
+    # coverage report has paths starting .tox and azure
+    # coverage combine fixes this with the help of tox.ini[coverage:paths]
+    combine_coverage_files(targeted_packages)
 
     coverage_files = []
     # generate coverage files
@@ -196,6 +221,11 @@ def prep_and_run_tox(targeted_packages, parsed_args, options_array=[]):
 
         local_options_array = options_array[:]
 
+        # Get code coverage params for current package
+        package_name = os.path.basename(package_dir)
+        coverage_commands = create_code_coverage_params(parsed_args, package_name)
+        local_options_array.extend(coverage_commands) 
+
         # if we are targeting only packages that are management plane, it is a possibility
         # that no tests running is an acceptable situation
         # we explicitly handle this here.
@@ -208,6 +238,7 @@ def prep_and_run_tox(targeted_packages, parsed_args, options_array=[]):
             )
         ):
             local_options_array.append("--suppress-no-test-exit-code")
+
 
         # if not present, re-use base
         if not os.path.exists(destination_tox_ini) or (