+++ sdk/keyvault/azure-keyvault-certificates/samples/recover_purge_operations_async.py
@@ -4,6 +4,7 @@
 # ------------------------------------
 import asyncio
 import os
+from azure.keyvault.certificates import CertificatePolicy
 from azure.keyvault.certificates.aio import CertificateClient
 from azure.identity.aio import DefaultAzureCredential
 from azure.core.exceptions import HttpResponseError
@@ -45,11 +46,13 @@ async def run_sample():
         bank_cert_name = "BankRecoverCertificate"
         storage_cert_name = "ServerRecoverCertificate"
 
-        bank_certificate_poller = await client.create_certificate(name=bank_cert_name)
-        storage_certificate_poller = await client.create_certificate(name=storage_cert_name)
+        bank_certificate = await client.create_certificate(
+            name=bank_cert_name, policy=CertificatePolicy.get_default()
+        )
+        storage_certificate = await client.create_certificate(
+            name=storage_cert_name, policy=CertificatePolicy.get_default()
+        )
 
-        bank_certificate = await bank_certificate_poller
-        storage_certificate = await storage_certificate_poller
         print("Certificate with name '{0}' was created.".format(bank_certificate.name))
         print("Certificate with name '{0}' was created.".format(storage_certificate.name))